[TF-TRT] Base Test Split to avoid OOM on GPUs with limited memory
