+++ tensorflow/python/compiler/tensorrt/test/base_test.py
@@ -19,7 +19,6 @@ import numpy as np
 from tensorflow.python.compiler.tensorrt.test import tf_trt_integration_test_base as trt_test
 from tensorflow.python.framework import constant_op
 from tensorflow.python.framework import dtypes
-from tensorflow.python.framework import ops
 from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import math_ops
 from tensorflow.python.ops import nn
@@ -162,151 +161,6 @@ class SimpleMultiEnginesTest2(trt_test.TfTrtIntegrationTestBase):
         not (trt_test.IsQuantizationMode(run_params.precision_mode) and
              not run_params.use_calibration)), "test FP32 and non-calibration"
 
-
-class ConstInputTest(trt_test.TfTrtIntegrationTestBase):
-
-  def GraphFn(self, inp):
-    """Create a graph containing multiple segment."""
-    n = inp
-    c = constant_op.constant(1.0, name="c")
-    # Adds data dependency from the constant op to a trt incompatible op,
-    # and adds data dependency from the trt incompatible op to the other
-    # ops, to make sure the constant op cannot be contracted with any trt
-    # segment that depends on it.
-    n = self.trt_incompatible_binary_op(n, c, name="incompatible")
-    n = math_ops.add(n, c, name="add")
-    n = math_ops.mul(n, n, name="mul")
-    n = math_ops.add(n, n, name="add1")
-    n = self.trt_incompatible_op(n, name="incompatible1")
-    n = math_ops.add(n, c, name="add2")
-    n = math_ops.mul(n, n, name="mul1")
-    n = math_ops.add(n, n, name="add3")
-    return array_ops.squeeze(n, name="output_0")
-
-  def GetParams(self):
-    shapes = [[2, 32, 32, 3]]
-    return self.BuildParams(self.GraphFn, dtypes.float32, input_shapes=shapes,
-                            output_shapes=shapes)
-
-  def ExpectedEnginesToBuild(self, run_params):
-    """Return the expected engines to build."""
-    return {
-        "TRTEngineOp_000": ["add", "add1", "mul"],
-        "TRTEngineOp_001": ["add2", "add3", "mul1"]
-    }
-
-  def ExpectedConnections(self, run_params):
-    """Returns the expected edges."""
-    return {
-        "input_0": set(),
-        "c": set(),
-        "incompatible": {"input_0", "c"},
-        "TRTEngineOp_000": {"incompatible"},
-        "incompatible1": {"TRTEngineOp_000"},
-        "TRTEngineOp_001": {"incompatible1"},
-        "output_0": {"TRTEngineOp_001"},
-    }
-
-
-class ConstDataInputSingleEngineTest(trt_test.TfTrtIntegrationTestBase):
-
-  def GraphFn(self, inp):
-    """Create a graph containing single segment."""
-    n = inp
-    c = constant_op.constant(1.0, name="c")
-    n = math_ops.add(n, c, name="add")
-    n = math_ops.mul(n, n, name="mul")
-    n = math_ops.add(n, n, name="add1")
-    return array_ops.squeeze(n, name="output_0")
-
-  def GetParams(self):
-    shapes = [[2, 32, 32, 3]]
-    return self.BuildParams(self.GraphFn, dtypes.float32, input_shapes=shapes,
-                            output_shapes=shapes)
-
-  def ExpectedEnginesToBuild(self, run_params):
-    """Return the expected engines to build."""
-    return {"TRTEngineOp_000": ["c", "add", "add1", "mul"]}
-
-
-class ConstDataInputMultipleEnginesTest(trt_test.TfTrtIntegrationTestBase):
-
-  def GraphFn(self, inp):
-    """Create a graph containing multiple segment."""
-    n = inp
-    c = constant_op.constant(1.0, name="c")
-    n = math_ops.add(n, c, name="add")
-    n = math_ops.mul(n, n, name="mul")
-    n = math_ops.add(n, n, name="add1")
-    n = self.trt_incompatible_op(n, name="incompatible1")
-    n = math_ops.add(n, c, name="add2")
-    n = math_ops.mul(n, n, name="mul1")
-    n = math_ops.add(n, n, name="add3")
-    return array_ops.squeeze(n, name="output_0")
-
-  def GetParams(self):
-    shapes = [[2, 32, 32, 3]]
-    return self.BuildParams(self.GraphFn, dtypes.float32, input_shapes=shapes,
-                            output_shapes=shapes)
-
-  def ExpectedEnginesToBuild(self, run_params):
-    """Return the expected engines to build."""
-    return {
-        "TRTEngineOp_000": ["add2", "add3", "mul1"],
-        # Why segment ["add", "add1", "mul"] was assigned segment id 1
-        # instead of 0: the parent node of this segment is actually const
-        # node 'c', but it's removed later since it's const output of the
-        # segment which is not allowed.
-        "TRTEngineOp_001": ["add", "add1", "mul"]
-    }
-
-
-class ControlDependencyTest(trt_test.TfTrtIntegrationTestBase):
-
-  def GraphFn(self, inp):
-    """Create a graph containing multiple segments."""
-    c1 = constant_op.constant(1.0, name="c1")
-    c2 = constant_op.constant(2.0, name="c2")
-    d1 = self.trt_incompatible_op(inp, name="d1")
-    d2 = self.trt_incompatible_binary_op(inp, inp, name="d2")
-    with ops.control_dependencies([d1]):
-      add = math_ops.add(inp, c1, name="add")
-    mul = math_ops.mul(add, add, name="mul")
-    add1 = math_ops.add(mul, mul, name="add1")
-    edge = self.trt_incompatible_op(add1, name="incompatible")
-    with ops.control_dependencies([d1, d2, add1]):
-      add2 = math_ops.add(edge, c2, name="add2")
-    mul1 = math_ops.mul(add2, add2, name="mul1")
-    add3 = math_ops.add(mul1, mul1, name="add3")
-    inc1 = self.trt_incompatible_binary_op(d1, add3, name="incompatible1")
-    inc2 = self.trt_incompatible_binary_op(d2, inc1, name="incompatible2")
-    return array_ops.squeeze(inc2, name="output_0")
-
-  def GetParams(self):
-    shapes = [[2, 32, 32, 3]]
-    return self.BuildParams(self.GraphFn, dtypes.float32, input_shapes=shapes,
-                            output_shapes=shapes)
-
-  def ExpectedEnginesToBuild(self, run_params):
-    """Return the expected engines to build."""
-    return {
-        "TRTEngineOp_000": ["c1", "add", "add1", "mul"],
-        "TRTEngineOp_001": ["c2", "add2", "add3", "mul1"]
-    }
-
-  def ExpectedConnections(self, run_params):
-    """Returns the expected edges."""
-    return {
-        "input_0": set(),
-        "d1": {"input_0"},
-        "d2": {"input_0"},
-        "TRTEngineOp_000": {"input_0", "^d1"},
-        "incompatible": {"TRTEngineOp_000"},
-        "TRTEngineOp_001": {"incompatible", "^d2"},
-        "incompatible1": {"TRTEngineOp_001", "d1"},
-        "incompatible2": {"incompatible1", "d2"},
-        "output_0": {"incompatible2"},
-    }
 
 if __name__ == "__main__":
   test.main()
